Convert to component architecture
